requirements.txt
streamlit>=1.33.0
supabase>=2.4.2
python-dotenv>=1.0.1
pandas>=2.0.0
plotly>=5.18.0
